cmd: use stdout fd to check whether we're in a terminal
This will make tsuru commands a bit more 'grep friendly' since we won't
wrap tables to the terminal size.

If the env TSURU_FORCE_WRAP is set with a non-empty value, tsuru will
revert to the old behavior wrapping even if output is not a terminal,
thus making tsuru 'lolcat friendly'.